move loadMode to text editor
